lift-simulator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.27.0
pandas
requests
python-multipart
openpyxl
pydantic>=2.4.0  
python-dotenv

